power.h tested

+++ src/Device.h
@@ -104,8 +104,8 @@ public:
                          double G_coeff, double high_G, double low_G, std::vector<std::string> metals);
 
     // update the power of each site
-    std::map<std::string, int> updatePower(int num_atoms_first_layer, double Vd, double high_G, double low_G_1,
-                                           std::vector<std::string> metals, double m_e, double V0);
+    std::map<std::string, double> updatePower(int num_atoms_first_layer, double Vd, double high_G, double low_G_1,
+                                              std::vector<std::string> metals, double m_e, double V0);
 
     // update the temperature of each site
     void updateTemperature();
